requirements.txt
fastapi
uvicorn
requests
line-bot-sdk
python-dotenv
